+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - Added support for the `backgroundcolor` property for Tiled maps.
+- Added `OnCrash` event to `SceneMgr` which allows to catch and recover from the exceptions within a scene.
 
 ### Fixed: